pyproject.toml
[tool.poetry]
name = "fixed-income-annuity"
version = "0.1.0"
description = ""
authors = ["pathikrit <pathikritbhowmick@msn.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
